update preprocess scripts.

+++ pipelines/calamity/stage_data/task_scripts/do_DOWNLOAD.sh
@@ -10,15 +10,18 @@ secrets_dir="${4}"
 gps=$(get_gps $fn)
 
 # copy settings.yaml and secrets to current dir
-cp ${secrets_dir}/settings.yaml ./
-cp ${secrets_dir}/client_secrets.json ./
+if [ ! -f "client_secrets.json" ]
+then
+  cp ${secrets_dir}/client_secrets.json ./
+fi
+if [ ! -f "settings.yaml" ]
+then
+  cp ${secrets_dir}/settings.yaml ./
+fi
+
 chmod g-rwx settings.yaml
 chmod o-rwx client_secrets.json
 
 
 echo mwa_download_gdrive.py --data_folder ${data_folder} --cal_folder ${cal_folder} --gpstime ${gps}
 mwa_download_gdrive.py --data_folder ${data_folder} --cal_folder ${cal_folder} --gpstime ${gps}
-
-# remove secrets and settings.
-rm -rf client_secrets.json
-rm -rf settings.yaml
